Rename: register -> signup
